+++ tests/schema/models_schema_create.py
@@ -81,6 +81,7 @@ class SourceFields(Model):
 
     class Meta:
         table = "sometable"
+        indexes = [["chars"]]
 
 
 class DefaultPK(Model):